requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.96.0
anyio==4.13.0
asttokens==3.0.1
async-timeout==4.0.3
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
executing==2.2.1
frozenlist==1.8.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
ipykernel==7.2.0
ipython==8.39.0
jedi==0.19.2
jiter==0.14.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.15
langchain-anthropic==1.4.1
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.0
langchain-openai==1.2.0
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.2
langdetect==1.0.9
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.10
langgraph-sdk==0.3.13
langsmith==0.7.33
lz4==4.4.5
marshmallow==3.26.2
matplotlib-inline==0.2.1
-e git+https://github.com/FelipeErmeson/model-qa-eval.git@8a5d430fe1a33f4afbccd241ea53fb65ac34b852#egg=model_qa_eval
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.6
protoc-gen-openapiv2==0.0.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==24.0.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
scikit-learn==1.7.2
scipy==1.15.3
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
stack-data==0.6.3
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
